# src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  title: core-product
  description: Products
  termsOfService: https://swagger.io/terms/
  contact:
    email: villavek1@gmail.com
  version: 1.0.0
...
servers:
  - url: http://localhost:8080/core-product/v1
...
paths:
...
  /products/{id}:
    get:
      tags:
        - Product
      summary: Allows you to obtain a product
      operationId: findById
      parameters:
        - $ref: '#/components/parameters/repositoryTypeCode'
        - $ref: '#/components/parameters/id'
      responses:
        204:
          description: The resource was send successfully.
        200:
          description: The resource was send successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        400:
          description: Invalid data supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
    put:
      tags:
        - Product
      summary: Allows you to update a product
      operationId: update
      parameters:
        - $ref: '#/components/parameters/repositoryTypeCode'
        - $ref: '#/components/parameters/id'
      requestBody:
        description: Product object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductRequest'
      responses:
        204:
          description: The resource was send successfully.
        400:
          description: Invalid data supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
        409:
          description: Invalid data supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
    delete:
      tags:
        - Product
      summary: Allows you to delete a product
      operationId: delete
      parameters:
        - $ref: '#/components/parameters/repositoryTypeCode'
        - $ref: '#/components/parameters/id'
      responses:
        204:
          description: The resource was send successfully.
        400:
          description: Invalid data supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
        409:
          description: Invalid data supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
security:
  - bearerAuth: [ ]
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    repositoryTypeCode:
      name: repositoryTypeCode
      in: header
      required: true
      description: Repository type code
      schema:
        type: string
        example: mongodb
        pattern: ^(mongodb|mysql)$
    id:
      name: id
      in: path
      required: true
      description: ID of the product to update
      schema:
        type: string
        example: 07c796f9-da9f-43aa-a8ae-fbd9ca24bc24
        pattern: ^[A-Za-z0-9-]+$
        minLength: 36
        maxLength: 36
  schemas:
    ProductRequest:
      type: object
      required:
        - code
        - name
      properties:
        code:
          type: string
          pattern: ^[A-Z0-9]+$
          minLength: 4
          maxLength: 4
          example: P001
        name:
          type: string
          pattern: ^[A-Za-z0-9 ]+$
          minLength: 4
          maxLength: 4
          example: Seguro de Vida
...
    ProductResponse:
      type: object
      properties:
        id:
          type: string
          example: 07c796f9-da9f-43aa-a8ae-fbd9ca24bc24
        code:
          type: string
          example: P001
        name:
          type: string
          example: Seguro de Vida
    ApiException:
      title: ApiException
      type: object
      properties:
        code:
          type: string
          description: Codigo de error de Sistema
          example: TL0001
        description:
          type: string
          description: Descripcion del error de Sistema
          example: Error al llamar al servicio
        errorType:
          type: string
          description: Tipo de Error de Sistema
          example: TECHNICAL
        exceptionDetails:
          type: array
          description: Lista de detalles del error
          items:
            $ref: '#/components/schemas/ApiExceptionDetail'
        properties:
          type: object
          properties: { }
          description: Lista de propiedades especiales.
      description: Datos del error de sistema.
    ApiExceptionDetail:
      title: ApiExceptionDetail
      type: object
      properties:
        code:
          type: string
          description: Codigo de error del Detalle/Proveedor
          example: MB0008
        component:
          type: string
          description: Nombre del componente de falla
          example: MB.CardInqV2
        description:
          type: string
          description: Descripcion del Detalle
          example: Codigo invalido para el canal
      description: Datos del error técnico.